Delete lcd_backlight_hal.c
The functionality can now be found in the infinity_erogodox.c file
instead.
